explanation behind the addition of parameter methods
this includes the explanation of the hierarchial order (Neuron -> Layer -> MLP) to aggregate parameters at each level of the hierarchy. also includes the logical order to the calculation of parameters (where n.parameters() starts, how each neuron's parameters is ordered by the number of its weights and its bias, and ALSO why the inputs (input layer) are not collected while collecting the parameters.